implement emqx interface methods for cloud and on-prem
