+++ mq/mq.go
@@ -43,19 +43,19 @@ func SetupMQTT() {
 	if servercfg.GetBrokerType() == servercfg.EmqxBrokerType {
 		time.Sleep(10 * time.Second) // wait for the REST endpoint to be ready
 		// setup authenticator and create admin user
-		if err := CreateEmqxDefaultAuthenticator(); err != nil {
+		if err := emqx.CreateEmqxDefaultAuthenticator(); err != nil {
 			logger.Log(0, err.Error())
 		}
-		DeleteEmqxUser(servercfg.GetMqUserName())
-		if err := CreateEmqxUser(servercfg.GetMqUserName(), servercfg.GetMqPassword(), true); err != nil {
+		emqx.DeleteEmqxUser(servercfg.GetMqUserName())
+		if err := emqx.CreateEmqxUser(servercfg.GetMqUserName(), servercfg.GetMqPassword(), true); err != nil {
 			log.Fatal(err)
 		}
 		// create an ACL authorization source for the built in EMQX MNESIA database
-		if err := CreateEmqxDefaultAuthorizer(); err != nil {
+		if err := emqx.CreateEmqxDefaultAuthorizer(); err != nil {
 			logger.Log(0, err.Error())
 		}
 		// create a default deny ACL to all topics for all users
-		if err := CreateDefaultDenyRule(); err != nil {
+		if err := emqx.CreateDefaultDenyRule(); err != nil {
 			log.Fatal(err)
 		}
 	}
